+++ src/server/controllers/sendChangePasswordEmail.ts
@@ -40,19 +40,14 @@ import {
 } from '@/server/lib/okta/idx/identify';
 import {
 	challenge,
-	isChallengeAnswerCompleteLoginResponse,
-	validateChallengeAnswerRemediation,
 	validateChallengeRemediation,
 } from '@/server/lib/okta/idx/challenge';
 import { recover } from '@/server/lib/okta/idx/recover';
 import { findAuthenticatorId } from '@/server/lib/okta/idx/shared/findAuthenticatorId';
-import { submitPassword } from '@/server/lib/okta/idx/shared/submitPasscode';
-import { credentialEnroll } from '@/server/lib/okta/idx/credential';
 import {
-	resetPassword,
-	validateRecoveryToken,
-} from '@/server/lib/okta/api/authentication';
-import { validateEmailAndPasswordSetSecurely } from '@/server/lib/okta/validateEmail';
+	forceUserIntoActiveState,
+	sendVerifyEmailAuthenticatorIdx,
+} from '@/server/controllers/oktaIdxShared';
 
 const { passcodesEnabled } = getConfiguration();
 
@@ -246,118 +241,22 @@ export const changePasswordEmailIdx = async ({
 						addQueryParamsToPath(emailSentPage, res.locals.queryParams),
 					);
 				} else if (passwordAuthenticatorId && !emailAuthenticatorId) {
-					// user has only password authenticator so:
+					// user has only password authenticator so they're in the
 					// 2. ACTIVE users - has only password authenticator (okta idx email not verified)
-
-					// If the user only has a password authenticator, then they failed to use a passcode
-					// to verify their account when they created it, and instead set a password using
-					// the okta classic reset password flow. In this case, we have to have to enroll
-					// the user in the email authenticator to allow them to reset their password.
-
-					// We do this by first setting a placeholder password for the user, then using the
-					// identify flow to authenticate the user with the placeholder password. After
-					// authenticating the user, the IDX API will tell us that the user needs to enroll
-					// in the email authenticator, which we can then use to send them a passcode to verify
-					// their account.
-					// Once they've verified their account, we use the okta classic api to generate a
-					// recover token, and instantly show the set password page to the user for them to
-					// set their password.
-
-					// set a placeholder password for the user
-					const placeholderPassword = await dangerouslySetPlaceholderPassword({
-						id: user.id,
-						ip: req.ip,
-						returnPlaceholderPassword: true,
+					// We need to send these users a verification email with passcode to verify their
+					// email before they can reset their password.
+					// See the method documentation for additional context
+					await sendVerifyEmailAuthenticatorIdx({
+						user,
+						identifyResponse,
+						passwordAuthenticatorId,
+						req,
+						res,
 					});
-
-					// call "challenge" to start the password authentication process
-					const challengePasswordResponse = await challenge(
-						introspectResponse.stateHandle,
-						{
-							id: passwordAuthenticatorId,
-							methodType: 'password',
-						},
-						req.ip,
-					);
-
-					// validate that the response from the challenge endpoint is a password authenticator
-					// and has the "recover" remediation
-					validateChallengeRemediation(
-						challengePasswordResponse,
-						'challenge-authenticator',
-						'password',
-					);
-
-					// call "challenge/answer" to answer the password challenge
-					const challengeAnswerResponse = await submitPassword({
-						password: placeholderPassword,
-						stateHandle: challengePasswordResponse.stateHandle,
-						introspectRemediation: 'challenge-authenticator',
-						ip: req.ip,
-					});
-
-					// check the response from the challenge/answer endpoint
-					// if it's a "CompleteLoginResponse" then Okta is in the state
-					// where email verification or enrollment is disabled, and a user
-					// can authenticate with a password only
-					// in this case we want to fall back to the classic reset password flow
-					// as this is the only way for these users to reset their password while
-					// Okta is in this state
-					if (isChallengeAnswerCompleteLoginResponse(challengeAnswerResponse)) {
-						// track the metric so we can see if we accidentally hit this case
-						trackMetric('OktaIDXEmailVerificationDisabled');
-						// throw an error to fall back to the classic reset password flow
-						throw new OktaError({
-							message: `Okta changePasswordEmailIdx failed as email verification or enrollment is disabled in Okta`,
-						});
-					}
-
-					// otherwise the response is a "ChallengeAnswerResponse" and we can continue
-					// but first we have to check that the response remediation is the "select-authenticator-enroll"
-					validateChallengeAnswerRemediation(
-						challengeAnswerResponse,
-						'select-authenticator-enroll',
-					);
-
-					// check for the email authenticator id in the response to make sure that it's the correct enrollment flow
-					const challengeAnswerEmailAuthenticatorId = findAuthenticatorId({
-						authenticator: 'email',
-						response: challengeAnswerResponse,
-						remediationName: 'select-authenticator-enroll',
-					});
-
-					// if the email authenticator id is not found, then throw an error to fall back to the classic reset password flow
-					if (!challengeAnswerEmailAuthenticatorId) {
-						throw new OktaError({
-							message: `Okta changePasswordEmailIdx failed as email authenticator id is not found in the response`,
-						});
-					}
-
-					// call the "challenge" endpoint to start the email challenge process
-					// and send the user a passcode
-					const challengeEmailResponse = await credentialEnroll(
-						challengeAnswerResponse.stateHandle,
-						{
-							id: challengeAnswerEmailAuthenticatorId,
-							methodType: 'email',
-						},
-						req.ip,
-					);
 
 					// track the success metrics
 					trackMetric('OktaIDXResetPasswordSend::Success');
 					trackMetric(`OktaIDXResetPasswordSend::${user.status}::Success`);
-
-					// at this point the user will have been sent an email with a passcode
-
-					// set the encrypted state cookie to persist the email and stateHandle
-					// which we need to persist during the passcode reset flow
-					setEncryptedStateCookie(res, {
-						email: user.profile.email,
-						stateHandle: challengeEmailResponse.stateHandle,
-						stateHandleExpiresAt: challengeEmailResponse.expiresAt,
-						userState: 'ACTIVE_PASSWORD_ONLY',
-					});
 
 					// show the email sent page, with passcode instructions
 					return res.redirect(
@@ -413,77 +312,20 @@ export const changePasswordEmailIdx = async ({
 					});
 				}
 
-				// 1. deactivate the user
 				try {
-					await deactivateUser({
-						id: user.id,
-						ip: req.ip,
-					});
-					trackMetric('OktaDeactivateUser::Success');
-				} catch (error) {
-					trackMetric('OktaDeactivateUser::Failure');
-					logger.error(
-						'Okta user deactivation failed',
-						error instanceof OktaError ? error.message : error,
-					);
-					throw error;
-				}
-
-				// 2. activate the user
-				try {
-					const tokenResponse = await activateUser({
-						id: user.id,
-						ip: req.ip,
-					});
-					if (!tokenResponse?.token.length) {
-						throw new OktaError({
-							message: `Okta user activation failed: missing activation token`,
-						});
-					}
-
-					// 3. use the recovery token to set a placeholder password for the user
-					// Validate the token
-					const { stateToken } = await validateRecoveryToken({
-						recoveryToken: tokenResponse.token,
-						ip: req.ip,
-					});
-					// Check if state token is defined
-					if (!stateToken) {
-						throw new OktaError({
-							message:
-								'Okta set placeholder password failed: state token is undefined',
-						});
-					}
-					// Set the placeholder password as a cryptographically secure UUID
-					const placeholderPassword = crypto.randomUUID();
-					await resetPassword(
-						{
-							stateToken,
-							newPassword: placeholderPassword,
-						},
-						req.ip,
-					);
-
-					// Unset the emailValidated and passwordSetSecurely flags
-					await validateEmailAndPasswordSetSecurely({
-						id: user.id,
-						ip: req.ip,
-						flagStatus: false,
-					});
-
-					// track the success metrics
-					trackMetric(
-						`OktaIDXResetPasswordSend::${user.status as Status}::Success`,
-					);
-
-					// now that the placeholder password has been set, the user will be in
+					// force the user into the ACTIVE state
+					// either
 					// 1. ACTIVE users - has email + password authenticator (okta idx email verified)
-					// or 2. ACTIVE users - has only password authenticator (okta idx email not verified)
-					// so we can call this method again to send the user a passcode
-					// They can't be in the 3. ACTIVE users - has only email authenticator (SOCIAL users, no password)
-					// as we've just set a placeholder password for them
-					// but we first need to get the updated user object
+					// 2. ACTIVE users - has only password authenticator (okta idx email not verified)
+					await forceUserIntoActiveState({
+						req,
+						user,
+					});
+
+					// get the updated user to get the correct status
 					const updatedUser = await getUser(user.id, req.ip);
+
+					// call this method again to send the user a passcode as they'll now be in one of the ACTIVE states
 					return changePasswordEmailIdx({
 						req,
 						res,
